Downloads/lasalle-spelling-bee-main (2)/experimental-spellingbeee/requirements.txt
flask==2.3.3
edge-tts==6.1.9
gTTS==2.5.1
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.2.0
gunicorn==21.2.0
flask-socketio==5.6.1
